requirements.txt
matplotlib==3.7.0
NBT==1.5.1
numpy==1.24.2
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.1
termcolor==2.2.0
opencv-python~=4.7.0.68
setuptools~=65.6.3
pyGLM==2.7.0
